Renamed MosaicConfigV2 -> MosaicConfig
